+++ mutant-lib/src/storage/mod.rs
@@ -1,7 +1,0 @@
-pub mod error;
-pub mod manager;
-
-pub use error::StorageError;
-
-#[cfg(test)]
-pub mod integration_tests;